refactor: move sect module
