+++ docs/module-structure.md
@@ -10,6 +10,6 @@ The project organizes game logic under the `game/` directory.
 - **game/debug.js** – wiring for optional debug controls.
 - **game/constants.js** – collection of shared gameplay constants used across systems.
 - **game/tooltip.js** – simple helpers for showing and hiding the UI tooltip.
-- **game/sect.js** – tracks discovered locations and holds colony tick/render stubs.
+- **game/sect.js** – implements the sect management system including constructs and colony helpers.
 
 The top-level `script.js` now merely orchestrates initialization, delegating all logic to the modules above.
